corrige betty problème

+++ singly_linked_lists/0-print_list.c
@@ -2,11 +2,15 @@
 #include <stdlib.h>
 #include "lists.h"
 /**
+ * print_list -  prints all the elements of a list
+ * @h: name of the first pointeur
  *
+ * Return: the number of nodes
 */
 size_t print_list(const list_t *h)
 {
-	size_t number_of_nodes = 0;
+	unsigned int number_of_nodes = 0;
+
 	while (h != NULL)
 	{
 		if (h->str == NULL)
